requirements.txt
#adafruit_dht
#atexit
#board
#python-dev
#python-rpi.gpio
paho-mqtt
psycopg2

